+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add service priority as a tag in opsgenie alerts.
 
+### Changed
+
+- Change CAPI version from v1alpha3 to v1beta1.
+
 ## [4.3.0] - 2022-08-02
 
 ### Fixed